app/requirements.txt
opencv-python
pandas
numpy
pillow
openpyxl

